optimasi foto ppr

+++ Modules/Ppr/Http/Controllers/PprProposalController.php
@@ -405,18 +405,22 @@ class PprProposalController extends Controller
                     'form_pribadi_keluarga_terdekat_no_telp' => $request->form_pribadi_keluarga_terdekat_no_telp,
                 ]);
 
-            foreach ($request->foto as $key => $value) {
-                if ($value['foto']) {
-                    Storage::delete($value['foto_lama']);
-                    $foto = $value['foto']->store('foto-ppr-pembiayaan');
+            if (request('perbarui_foto_pemohon') == 'Ya') {
+                foreach ($request->foto as $key => $value) {
+                    if ($value['foto']) {
+                        Storage::delete($value['foto_lama']);
+                        $foto = $value['foto']->store('foto-ppr-pembiayaan');
 
-                    FormPprFoto::where('form_ppr_pembiayaan_id', $id)->where('id', $value['id'])->update([
-                        'form_ppr_pembiayaan_id' => $id,
-                        'kategori' => $value['kategori'],
-                        'foto' => $foto,
-                    ]);
+                        FormPprFoto::where('form_ppr_pembiayaan_id', $id)->where('id', $value['id'])->update([
+                            'form_ppr_pembiayaan_id' => $id,
+                            'kategori' => $value['kategori'],
+                            'foto' => $foto,
+                        ]);
+                    }
                 }
+            } else {
             }
+
 
             FormPprDataPekerjaan::where('form_ppr_data_pribadi_id', $id)
                 ->update([
